chore: `has-*` shouldn't be experimental
